Pull out code to average and insert into circular buffers into utility.hpp

+++ RocketFlightComputer/src/states/01-Launch.h
@@ -12,6 +12,6 @@ public:
 private:
     bool motorBurnout = false;
     float accelerationBuffer[10] = {0};
-    int bufferIndex = 0;
+    size_t bufferIndex = 0;
     Debouncer motorBurnoutDebouncer = Debouncer(30);
 };
